+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - PR #5303 Add slice_strings functionality using delimiter string
 - PR #5394 Enable cast and binops with duration types (builds on PR 5359)
 - PR #5301 Add Java bindings for `zfill`
+- PR #5411 Enable metadata collection for chunked parquet writer
 - PR #5359 Add duration types
 - PR #5364 Validate array interface during buffer construction
 - PR #5418 Add support for `DataFrame.info`